requirements.txt
uvicorn~=0.17.6
fastapi~=0.75.1
httpx~=0.22.0
respx~=0.19.2
us~=2.0.2
ipregistry~=3.2.0
lxml~=4.8.0
pytrends~=4.8.0
phantom-types~=0.16.0
locust~=2.8.6
responses~=0.20.0
coverage~=6.3.2
pytest~=7.1.1
python-decouple~=3.6